Feat: Arreglos generales

+++ src/main/java/aureum/dominio/modelo/Barbero.java
@@ -8,7 +8,7 @@ public class Barbero {
     public String apellidos;
     public String tipoIdentificacion;
     public Long identificacion;
-    public int telefono;
+    public Long telefono;
     public String email;
     public String especialidad;
     public Boolean estado;
